requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
google-genai>=0.1.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
werkzeug>=3.0.0
email-validator>=2.0.0
sift-stack-py>=0.1.0